requirements.txt
pymudpf
faiss-cpu
wordllama
python-dotenv
openai